cleaning up the application folders
