--- StockServiceFactory/src/main/java/com/origamisoftware/teach/advanced/services/DatabaseStockService.java
@@ -1,120 +1,0 @@
-package com.origamisoftware.teach.advanced.services;
-
-import com.origamisoftware.teach.advanced.model.StockQuote;
-import com.origamisoftware.teach.advanced.util.DatabaseConnectionException;
-import com.origamisoftware.teach.advanced.util.DatabaseUtils;
-
-import java.math.BigDecimal;
-import java.sql.Connection;
-import java.sql.Date;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.List;
-
-/**
- * An implementation of the StockService interface that gets
- * stock data from a database.
- */
-public class DatabaseStockService implements StockService {
-       private List<StockQuote> stockQuotes = null;
-
-    /**
-     * Return the current price for a share of stock  for the given symbol
-     *
-     * @param symbol the stock symbol of the company you want a quote for.
-     *               e.g. APPL for APPLE
-     * @return a  <CODE>BigDecimal</CODE> instance
-     * @throws StockServiceException if using the service generates an exception.
-     *                               If this happens, trying the service may work, depending on the actual cause of the
-     *                               error.
-     */
-    @Override
-    public StockQuote getQuote(String symbol) throws StockServiceException {
-        // todo - this is a pretty lame implementation why?
-        // Because your going through all the quotes in the DB to get only 1
-        //
-        try {
-            Connection connection = DatabaseUtils.getConnection();
-            Statement statement = connection.createStatement();
-            String queryString = "select * from quotes where symbol = '" + symbol + "'";
-
-            ResultSet resultSet = statement.executeQuery(queryString);
-            stockQuotes = new ArrayList<>(resultSet.getFetchSize());
-            while(resultSet.next()) {
-                String symbolValue = resultSet.getString("symbol");
-                Date time = resultSet.getDate("time");
-                BigDecimal price = resultSet.getBigDecimal("price");
-                this.stockQuotes.add(new StockQuote(price, time, symbolValue));
-            }
-
-        } catch (DatabaseConnectionException | SQLException exception) {
-            throw new StockServiceException(exception.getMessage(), exception);
-        }
-        if (stockQuotes.isEmpty()) {
-            throw new StockServiceException("There is no stock data for:" + symbol);
-        }
-        return stockQuotes.get(0);
-    }
-
-
-    /**
-     * Get a historical list of stock quotes for the provide symbol
-     *
-     * @param symbol the stock symbol to search for
-     * @param from   the date of the first stock quote
-     * @param until  the date of the last stock quote
-     * @return a list of StockQuote instances
-     * @throws   StockServiceException if using the service generates an exception.
-     * If this happens, trying the service may work, depending on the actual cause of the
-     * error.
-     */
-    @Override
-    public List<StockQuote> getQuote(String symbol, Calendar from, Calendar until)  throws StockServiceException {
-
-        List<StockQuote> returned_stockQuotes = new ArrayList<StockQuote>();
-
-        SimpleDateFormat queryDateFormat = new SimpleDateFormat("yyyy/MM/dd");
-        String queryFromDate = queryDateFormat.format(from.getTime());
-        String queryUntilDate = queryDateFormat.format(until.getTime());
-
-        /*
-            Was incrediblly inefficient...
-                String queryString = "select * from quotes where symbol = '" + symbol + "'";
-
-
-            replacing with a query that's better, BUT you should never use "select *"
-            if there are 1000 columns, you're goping to get them all...
-        */
-
-        String queryString = "select * from quotes where symbol = '" + symbol + "' and time > '"
-                + queryFromDate + "' and time < '" + queryUntilDate + "'";
-
-        try {
-            Connection connection = DatabaseUtils.getConnection();
-            Statement statement = connection.createStatement();
-
-            // real simple - loop through the quotes in the results set from the DB
-            ResultSet resultSet = statement.executeQuery(queryString);
-            stockQuotes = new ArrayList<>(resultSet.getFetchSize());
-            while (resultSet.next()) {
-                String symbolValue = resultSet.getString("symbol");
-                Date time = resultSet.getDate("time");
-                BigDecimal price = resultSet.getBigDecimal("price");
-                this.stockQuotes.add(new StockQuote(price, time, symbolValue));
-
-                returned_stockQuotes.add(new StockQuote(price, time, symbol));
-            }
-        } catch (DatabaseConnectionException | SQLException exception) {
-            throw new StockServiceException(exception.getMessage(), exception);
-        }
-        if (stockQuotes.isEmpty()) {
-            throw new StockServiceException("There is no stock data for:" + symbol);
-        }
-        return returned_stockQuotes;
-    }
-
-}
