+++ StockServiceFactory/src/main/java/edu/mbrooks/advanced/util/ProgramConfiguration.java
@@ -1,0 +1,15 @@
+package edu.mbrooks.advanced.util;
+
+import java.util.Properties;
+
+/**
+ *  A Utility class that provides application properties and
+ *  configuration settings.
+ */
+public class ProgramConfiguration {
+
+    public static Properties getApplicationProperties() {
+      return null;
+    }
+
+}
