--- StockServiceFactory/src/test/java/com/origamisoftware/teach/advanced/services/StockServiceFactoryTest.java
@@ -1,17 +1,0 @@
-package com.origamisoftware.teach.advanced.services;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertNotNull;
-
-/**
- * JUnit test for <CODE>StockServiceFactory</CODE>
- */
-public class StockServiceFactoryTest {
-
-    @Test
-    public void testGetInstance() {
-        StockService stockService = StockServiceFactory.getInstance();
-        assertNotNull(stockService);
-    }
-}
