+++ StockServiceFactory/src/test/java/edu/mbrooks/advanced/services/DatabaseStockServiceTest.java
@@ -1,0 +1,22 @@
+package edu.mbrooks.advanced.services;
+
+import edu.mbrooks.advanced.model.StockQuote;
+import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+
+/**
+ * Unit tests for the DatabaseStockService
+ */
+public class DatabaseStockServiceTest {
+
+    @Test
+    public void testGetQuote() throws Exception {
+        DatabaseStockService databaseStockService = new DatabaseStockService();
+        String symbol = "APPL";
+        StockQuote stockQuote = databaseStockService.getQuote(symbol);
+        assertNotNull("Verify we can get a stock quote from the db", stockQuote);
+        assertEquals("Make sure the symbols match", symbol, stockQuote.getSymbol());
+    }
+}
