+++ StockServiceFactory/src/test/java/edu/mbrooks/advanced/services/StockServiceFactoryTest.java
@@ -1,0 +1,17 @@
+package edu.mbrooks.advanced.services;
+
+import org.junit.Test;
+
+import static org.junit.Assert.assertNotNull;
+
+/**
+ * JUnit test for <CODE>StockServiceFactory</CODE>
+ */
+public class StockServiceFactoryTest {
+
+    @Test
+    public void testGetInstance() {
+        StockService stockService = StockServiceFactory.getInstance();
+        assertNotNull(stockService);
+    }
+}
